requirements.txt
flask
gunicorn
ccxt
line-bot-sdk